+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.1.3] - Unreleased
+
+### Added
+
+- Add `csv::load` module
+  - New methods wrap over the errors of deser/parse steps
+
+### Changed
+
+- Moved templates data for tests `from tests::engines::{hfbi, niseci}::full` to `tests::test_utils`
+
 ## [0.1.2] - 2026-02-06
 
 ### Changed